Got rid of the "Static" parameter on all functions (because it's easier to just run a for loop not within the function, if anything I can reimplement it there).
Also I created a function that takes in a ggplot object and some latex script and converts the labels and title to latex. There's also a small example on how to use it.

Please note that with this package, you have to use two backslashes instead of one (i.e. do \\tau instead of \tau)